lab7-mikolaj-olesinski/requirements.txt
streamlit==1.40.2
langchain==0.3.27
langchain-google-genai==2.1.5
langchain-community==0.3.27
ddgs>=9.0.0
trafilatura==2.0.0
plotly==5.24.1
pandas==2.2.3
python-dotenv==1.0.1
redis==5.2.1
prometheus-client==0.21.0
